tools/export_onnx/requirements.txt
sentence-transformers>=3.0.0
optimum[exporters]>=1.17.0
onnxruntime>=1.16.0
numpy>=1.24.0
torch>=2.0.0
